contracts, test: add Validate library
