+++ contracts/DarknodeSlasher/DarknodeSlasher.sol
@@ -1,9 +1,8 @@
 pragma solidity ^0.5.8;
 
 import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
-import "openzeppelin-solidity/contracts/cryptography/ECDSA.sol";
 
-import "../libraries/String.sol";
+import "../libraries/Validate.sol";
 import "../DarknodeRegistry/DarknodeRegistry.sol";
 
 /// @notice DarknodeSlasher will become a voting system for darknodes to
@@ -70,7 +69,7 @@ contract DarknodeSlasher is Ownable {
         uint256 _validRound2,
         bytes calldata _signature2
     ) external {
-        address signer = validateDuplicatePropose(
+        address signer = Validate.duplicatePropose(
             _height,
             _round,
             _blockhash1,
@@ -83,47 +82,5 @@ contract DarknodeSlasher is Ownable {
         require(!slashed[_height][_round][signer], "already slashed");
         slashed[_height][_round][signer] = true;
         darknodeRegistry.slash(signer, msg.sender, maliciousSlashPercent);
-    }
-
-    function validateDuplicatePropose(
-        uint256 _height,
-        uint256 _round,
-        bytes memory _blockhash1,
-        uint256 _validRound1,
-        bytes memory _signature1,
-        bytes memory _blockhash2,
-        uint256 _validRound2,
-        bytes memory _signature2
-    ) public pure returns (address) {
-        require(_validRound1 != _validRound2, "same valid round");
-        address signer1 = recoverPropose(_height, _round, _blockhash1, _validRound1, _signature1);
-        address signer2 = recoverPropose(_height, _round, _blockhash2, _validRound2, _signature2);
-        require(signer1 == signer2, "different signer");
-        return signer1;
-    }
-
-    function recoverPropose(
-        uint256 _height,
-        uint256 _round,
-        bytes memory _blockhash,
-        uint256 _validRound,
-        bytes memory _signature
-    ) public pure returns (address) {
-        return ECDSA.recover(sha256(proposeMessage(_height, _round, _blockhash, _validRound)), _signature);
-    }
-
-    function proposeMessage(
-        uint256 _height,
-        uint256 _round,
-        bytes memory _blockhash,
-        uint256 _validRound
-    ) public pure returns (bytes memory) {
-        return abi.encodePacked(
-            "Propose(Height=", String.fromUint(_height),
-            ",Round=", String.fromUint(_round),
-            ",BlockHash=", string(_blockhash),
-            ",ValidRound=", String.fromUint(_validRound),
-            ")"
-        );
     }
 }
